0.55 "material" system for spawn correct fx effect in correct places :)

+++ core/src/com/cosma/annihilation/Components/GateComponent.java
@@ -6,7 +6,7 @@ import com.cosma.annihilation.EntityEngine.core.Component;
 public class GateComponent implements Component {
     public boolean isOpen = false;
     public boolean isMoving = false;
-    public int moveDistance = 2;
+    public float moveDistance = 2;
     public Vector2 targetPosition = new Vector2();
     public String gateName;
 }
